+++ packages/toolpad-app/src/toolpadDataSources/function/types.ts
@@ -1,3 +1,5 @@
+import { LogEntry } from './Console';
+
 export interface FunctionConnectionParams {
   secrets?: [string, string][];
 }
@@ -14,46 +16,6 @@ export interface PrivateQuery<I, R> {
 export interface PrivateQueries {
   debug: PrivateQuery<FunctionQuery, {}>;
 }
-
-export interface LogRequest {
-  method: string;
-  url: string;
-  headers: [string, string][];
-}
-
-export interface LogResponse {
-  status: number;
-  statusText: string;
-  ok: boolean;
-  headers: [string, string][];
-  bodyUsed: boolean;
-  redirected: boolean;
-  type: Response['type'];
-  url: string;
-}
-
-export interface LogConsoleEntry {
-  timestamp: number;
-  level: string;
-  kind: 'console';
-  args: any[];
-}
-
-export interface LogRequestEntry {
-  timestamp: number;
-  kind: 'request';
-  id: string;
-  request: LogRequest;
-}
-
-export interface LogResponseEntry {
-  timestamp: number;
-  kind: 'response';
-  id: string;
-  response: LogResponse;
-}
-
-export type LogEntry = LogConsoleEntry | LogRequestEntry | LogResponseEntry;
 
 export interface FunctionResult {
   data: any;
